requirements.txt
exceptiongroup==1.2.2
iniconfig==2.0.0
line_profiler==4.2.0
llvmlite==0.44.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
PyYAML==6.0.2
tomli==2.2.1
